FIXME: feat: Re-add NativeScript inspector sources
These were removed in the upgrade to V8 10.3.22. Re-add them. Note that
the ObjC++ files need to be compiled with -fobjc-arc or else connecting
the inspector crashes.

TODO: There are probably some other V8 headers that need to be added
once we start porting the inspector to the new API.